+++ scripts/package_it.py
@@ -53,8 +53,8 @@ PACKAGES = [
         "suffix": "src",
         "root_files": [
             "Cargo.toml",
-            "CHANGELOG.md",
-            "README.md",
+            # "CHANGELOG.md",
+            # "README.md",
             # "ROADMAP.md",
             # "SECURITY.md",
         ],
@@ -69,17 +69,17 @@ PACKAGES = [
         "file_pattern": "*.rs",
         "description": "All integration tests",
     },
-    {
-        "suffix": "fuzz",
-        "root_files": [
-            "fuzz/Cargo.toml",
-            ".github/workflows/fuzz-nightly.yml",
-            ".github/workflows/fuzz-quick.yml",
-        ],
-        "include_dirs": ["fuzz/fuzz_targets"],
-        "file_pattern": "*.rs",
-        "description": "Fuzzing targets",
-    },
+    # {
+    #     "suffix": "fuzz",
+    #     "root_files": [
+    #         "fuzz/Cargo.toml",
+    #         ".github/workflows/fuzz-nightly.yml",
+    #         ".github/workflows/fuzz-quick.yml",
+    #     ],
+    #     "include_dirs": ["fuzz/fuzz_targets"],
+    #     "file_pattern": "*.rs",
+    #     "description": "Fuzzing targets",
+    # },
     # {
     #     "suffix": "mod",
     #     "root_files": [],